osquery/requirements.txt
kafka-python
pygtail